requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.6
django-crispy-forms==1.11.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
jsonfield==3.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
